Move functions to utils.py #2929
